refactor media progress
